+++ config.js
@@ -1,5 +1,5 @@
 window.APP_CONFIG = {
   appName: 'Mampfo',
   shortName: 'Mampfo',
-  version: '0.5.3'
+  version: '0.6.1'
 };